requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
pymongo==4.7.2  # or similar version
python-dotenv==1.0.1  # or your current version
